remove duplicate windows_helpers.cpp, refactor
